Backup D-Star Time Announcement setting
The Time Announcement setting is handled a bit differently, and
requires backing up the empty /etc/timeserver.dissable to keep
the service from running. Add this file to the backup archive
to keep it in sync with a restore operation, and remove it
during a restore to clear any existing setting.